chore(package): set up Travis-CI, Coveralls & semantic-release
Now that the repository is public, we will be using
[semantic-release](https://github.com/semantic-release/semantic-release) to manage our releases from
here on out. This meant we needed to set up Travis-CI and (optionally) Coveralls. It also meant we
could start removing a few relics of our previous build process on our firewalled Jenkins server. On
a happier note, this means there are more badges for us to add! :sunglasses: