feat(station-finder): Implement changes to support Station Finder Service v3
V3 of the Station Finder Service is close to being ready for use in production environments. This

commit implements the necessary changes (mostly done at a data modeling level) to support the new

version of the API. While there are no breaking changes to the code from a consumer perspective,

station searches may now return different data compared to what was returned by Station Finder v2,

which is why this is being treated as a new feature release.