fix(station): Do not fall back to small-logo if logo does not exist
In Station Finder v2, we had logic to fall back to the "small" station logo if the larger, NPR

One-preferred logo did not exist. However, this logic was never activated because Station Finder v2

wasn't returning the small logos. Station Finder v3 does correctly return the small logos now, but

we realized pretty quickly after testing that we *don't* want this fallback logic, as it results in

a suboptimal UI treatment.