chore(build): change node version to 4.x and up
Changing the minimum required version of Node in package.json to be 4.x or better. There'll be a

second release for 6.x or better. The reason for this is that many libraries could still support 4.x

since it is still under LTS per https://github.com/nodejs/Release. Versions being built in Travis

are 8, 6, and 4.

Closes #9