fix(api): Implement the new hostnames for NPR One API calls
IMPORTANT: As part of this change, the `apiBaseUrl` and `apiVersion` parameters on the config

parameters are no longer in use and will log a warning if they attempt to be set. Since most clients

should not have been setting custom values for these but relying on the defaults instead, we're not

considering this a breaking change. However, any clients requiring use of a custom environment

should now use the `subdomain` parameter that has been added to the config instead.