feat(authorization): Add support for other auth proxies
We have at least one other client that uses an auth proxy that works slightly differently than our

PHP-based one, eschewing the use of cookies in favor of handling refresh tokens and device codes

manually. (The former will be stored in localStorage much like the access tokens.) The SDK needs to

be able to gracefully handle these other use cases while not breaking the existing experience for

clients using the PHP-based proxy; this commit takes care of that. NOTE: I haven't had time to

augment the unit tests to cover all of the new cases, so test coverage will go down as a result of

this push, but I don't have any more time to work on this today, so I wanted to push what I have so

far. My hope is to wrap up the tests and get coverage back up to 100% tomorrow.

DIG-11363