build(node): upgrade to Node.js v10.x or higher
This package was still supporting Node.js v6, which has been EOL for a number of years now. I know
Node.js v10 is EOL fairly soon as well, but this functions partly as a test balloon to see how easy
upgrading Node is at this stage, given that this package has not really been maintained.

BREAKING CHANGE: Requires a minimum version of Node.js v10.13 or higher. Note that depending on the
context (e.g. if you're just trying to run the compiled JS in a browser), this may continue to be
compatible with older systems, but going forward, we will not be going out of our way to stay
compatible with Node versions that are EOL.

Closes #15. Closes #26.